changes commit

+++ Code/website/api.py
@@ -48,22 +48,28 @@ def createCategory():
         category = Category(categoryName=data["name"])
         db.session.add(category)
         db.session.commit()
-        return "<h1>Category added succesfully</h1><a href='/addcategory'>Add more Categories</a> <a href = '/'>Go to Home page</a>", 201
+        return redirect("/categories")
 # api for deleting a category
 
 
-@api.route('deleteCategory', methods=["DELETE"])
+@api.route('/deleteCategory', methods=["GET"])
 def deleteCategory():
-    data = request.get_json()
-    categories = list(Category.query.filter_by(categoryName=data["name"]))
+    data = request.args.get("id")
+    if (data is None):
+        return "Please enter category id in URL"
+    categories = list(Category.query.filter_by(categoryId=data))
     if (len(categories) == 0):
         return "Category does not exist", 500
-    res = Category.query.filter_by(
-        categoryName=data["name"])[0]
+    res = categories[0]
     db.session.delete(res)
     db.session.commit()
-    print(res)
-    return "Category successfully deleted", 200
+    db.session.close()
+    products = list(Product.query.filter_by(categoryId=data))
+    if (len(products) > 0):
+        db.session.delete(products)
+        db.session.commit()
+
+    return redirect("/categories")
 
 # endregion
 
@@ -90,11 +96,11 @@ def getProducts():
 def createProduct():
     data = request.get_json()
     categoryId = request.args.get("id", type=int)
-    categoryCount = len(Category.query.all())
-    if (categoryId is None):
-        return "Category Id is null. Please enter category id in url params", 400
-    if (categoryId > categoryCount):
-        return "Please select valid category", 400
+    categoryCount = Category.query.all()
+    for cat in categoryCount:
+        if (cat.categoryId != categoryId):
+            return "Category invalid, Please select valid category", 400
+
     product = Product(productName=data["name"], unit=data["unit"],
                       rate=data["rate"], quantity=data['quantity'], categoryId=categoryId)
     db.session.add(product)
